load settings from config folder instead of config file and move function into utils
