+++ scripts/get-shapefiles.py
@@ -5,19 +5,7 @@ import os
 import zipfile
 import requests
 import shutil
-from utils import JURISDICTION_NAMES, find_jurisdiction, ROOTDIR, setup_source
-import yaml
-
-
-def load_settings(config_dir: str):
-    with open(f"{config_dir}/settings.yml", "r") as f_in:
-        settings = yaml.safe_load(f_in.read())
-    settings["jurisdictions"] = dict()
-    for file in glob.glob(f"{config_dir}/jurisdictions/*.yml"):
-        with open(file, "r") as f:
-            jur_settings = yaml.safe_load(f.read())
-            settings["jurisdictions"][jur_settings["name"]] = dict(jur_settings)
-    return settings
+from utils import JURISDICTION_NAMES, find_jurisdiction, ROOTDIR, setup_source, load_settings
 
 
 def download_from_tiger(jurisdiction, year):
